feat #10: coordenadas calculadas desde admin-django
