+++ aiding/information/urls.py
@@ -3,6 +3,6 @@ from .views import ResourceView
 
 urlpatterns = [
     path('resources/', ResourceView.as_view()),
-    path('resources/<int:id>', ResourceView.as_view())
+    path('resources/<int:resource_id>', ResourceView.as_view())
 
 ]
